Change EcsColor to use floats

+++ include/graphics.h
@@ -16,10 +16,10 @@ typedef struct EcsRect {
 } EcsRect;
 
 typedef struct EcsColor {
-    uint8_t r;
-    uint8_t g;
-    uint8_t b;
-    uint8_t a;
+    float r;
+    float g;
+    float b;
+    float a;
 } EcsColor;
 
 typedef struct EcsCanvas2D {
